sgst-api/requirements.txt
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.120.0
h11==0.16.0
idna==3.11
limits==5.8.0
mysql-connector-python==9.5.0
packaging==26.0
passlib==1.7.4
pycparser==3.0
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
slowapi==0.1.9
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wrapt==2.1.1
